Add meta description for search result snippets
This will suggest a page description as an option for search engines to
use in results. Some searches will currently use text with cookie info
because this is the very first DOM content with text on the page.

Search engines will create snippets automatically from content, so this
meta description can’t be forced, but hopefully persuade as more
relevant infromation to display than “Cookie settings” 🍪